Configuring conda 2

+++ SetUpLinuxOnRemote.sh
@@ -60,10 +60,6 @@ fi
 # Quiet log in to servers
 touch $HOME/.hushlogin
 
-printf "\nSourcing and binding\n\n"
-. $HOME/.bashrc
-#bind -f  $HOME/.inputrc
-
 # Check if mybin exists if not create it
 if [[ ! -d "$HOME/mybin" ]]; then
   mkdir -p $HOME/mybin
@@ -71,22 +67,64 @@ fi
 
 # Installing miniconda according to user's preference
 if [[ ${CONDA} == "conda" ]]; then
-  printf "\nMiniconda's default path is going to be "$HOME/miniconda".\n"
-  #Download the latest shell script
-  wget -q https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -P $HOME/mybin/
+  printf "\nLet's rock miniconda!\n"
+  if [[ -d $HOME/miniconda3/bin ]]; then
+    printf "\nMiniconda already exists at "$HOME/miniconda3". If you want to delete and re-install miniconda3 run $0 conda_reinstall.\n* Aborting repeated miniconda3 installation. *\n"
+    exit 0
+  else
+    printf "\nMiniconda's default path is going to be "$HOME/miniconda".\n"
+    #Download the latest shell script
+    wget -q https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -P $HOME/mybin/
 
-  #Make the miniconda installation script executable
-  chmod +x $HOME/mybin/Miniconda3-latest-Linux-x86_64.sh
+    #Make the miniconda installation script executable
+    chmod +x $HOME/mybin/Miniconda3-latest-Linux-x86_64.sh
 
-  #Run miniconda installation script
-  $HOME/mybin/Miniconda3-latest-Linux-x86_64.sh -b -p $HOME/miniconda3
-  rm $HOME/mybin/Miniconda3-latest-Linux-x86_64.sh
+    #Run miniconda installation script
+    $HOME/mybin/Miniconda3-latest-Linux-x86_64.sh -b -p $HOME/miniconda3
+    rm $HOME/mybin/Miniconda3-latest-Linux-x86_64.sh
 
-  # Let's config and install some packages
-  conda config --add channels conda-forge;conda config --add channels defaults;conda config --add channels r;conda config --add channels bioconda
-  conda install samtools bedops parallel bcftools bedtools boto3 libopenblas
-  parallel --cite
+    # Let's config and install some packages
+    conda config --set auto_activate_base false
+    conda config --add channels conda-forge
+    conda config --add channels defaults
+    conda config --add channels r
+    conda config --add channels bioconda
+
+    printf "Installing basic packages: samtools bedops parallel bcftools bedtools boto3 libopenblas\n"
+    conda install -y samtools bedops parallel bcftools bedtools boto3 libopenblas
+
+    # Silencing parallel's warnings
+    parallel --cite
+  fi
+  elif [[ ${CONDA} == "conda_reinstall" ]]; then
+    printf "\nLet's rock miniconda!\n"
+    printf "Deleting miniconda3 from $HOME/miniconda3"
+    rm -rf $HOME/miniconda3
+    printf "\nMiniconda's new path is going to be "$HOME/miniconda3".\n"
+    #Download the latest shell script
+    wget -q https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -P $HOME/mybin/
+
+    #Make the miniconda installation script executable
+    chmod +x $HOME/mybin/Miniconda3-latest-Linux-x86_64.sh
+
+    #Run miniconda installation script
+    $HOME/mybin/Miniconda3-latest-Linux-x86_64.sh -b -p $HOME/miniconda3
+    rm $HOME/mybin/Miniconda3-latest-Linux-x86_64.sh
+
+    # Let's config and install some packages
+    conda config --set auto_activate_base false
+    conda config --add channels conda-forge
+    conda config --add channels defaults
+    conda config --add channels r
+    conda config --add channels bioconda
+
+    printf "Installing basic packages: samtools bedops parallel bcftools bedtools boto3 libopenblas\n"
+    conda install -y samtools bedops parallel bcftools bedtools boto3 libopenblas
 fi
+
+printf "\nSourcing .bashrc\n\n"
+. $HOME/.bashrc
+#bind -f  $HOME/.inputrc
 
 printf "All done! It's a good idea to log out and back in to enable .inputrc.\nSee you later, alligator!\n"
 exit 0
